net/tunstats: use high performance hash table implementation
Rather than using Go maps, which are optimized to be decent
for all operations: lookup, insertion, deletion, iteration, etc.

We only need to care about one operation: update.
An update is a lookup and insertion combined into one.
Futhermore, we have the benefit that statistics are reducible
such that they can be merged together from multiple entries.

We implement our own hash table, which uses open addressing
as opposed to chaining. Open addressing is simple and memory efficient,
but generally not as favored since deletion is more complicated and slower.
Fortunately, we don't care about deletion.

Also, use of open addressing makes it easier to just replace the table,
so that the table can be obtained with an atomic pointer load.
This allows concurrent goroutines to operate on different
entries in the same table.

Overall, this data structure is ~3-6x faster.
It is increasingly faster with more threads.
It uses significantly less memory since open addressing
represents the hash table as a single array.
No overhead is lost on pointer indirection and fragmentation.

Performance:

	name                           old time/op    new time/op    delta
	/SingleRoutine/SameConn-24        943µs ± 1%     291µs ± 0%  -69.19%  (p=0.008 n=5+5)
	/SingleRoutine/UniqueConns-24    3.05ms ± 1%    0.56ms ± 1%  -81.82%  (p=0.008 n=5+5)
	/MultiRoutine/SameConn-24        32.6ms ± 6%    13.6ms ± 1%  -58.22%  (p=0.008 n=5+5)
	/MultiRoutine/UniqueConns-24      131ms ± 5%      22ms ± 2%  -82.93%  (p=0.008 n=5+5)

	name                           old alloc/op   new alloc/op   delta
	/SingleRoutine/SameConn-24         960B ± 0%      456B ± 0%  -52.50%  (p=0.008 n=5+5)
	/SingleRoutine/UniqueConns-24    3.97MB ± 0%    0.53MB ± 0%  -86.76%  (p=0.008 n=5+5)
	/MultiRoutine/SameConn-24        3.00kB ±47%    2.34kB ±96%     ~     (p=0.206 n=5+5)
	/MultiRoutine/UniqueConns-24      127MB ± 0%      17MB ± 0%  -86.76%  (p=0.008 n=5+5)

	name                           old allocs/op  new allocs/op  delta
	/SingleRoutine/SameConn-24         3.00 ± 0%      3.00 ± 0%     ~     (all equal)
	/SingleRoutine/UniqueConns-24       276 ± 0%        28 ± 0%  -89.86%  (p=0.008 n=5+5)
	/MultiRoutine/SameConn-24          30.8 ±10%      29.6 ±18%     ~     (p=0.302 n=5+5)
	/MultiRoutine/UniqueConns-24      9.63k ± 1%     0.09k ± 1%  -99.07%  (p=0.008 n=5+5)

Optimizations:

	* Most packets have contiguous address and ports, which we can extract together.
	* Updates are mostly lock-free. There is a contended atomic integer update
	  when calling sync.RWMutex.RLock. This could be avoided in the future.
	* Map counts are represented using 32-bit since overflow is rare.
	  Overflow is handled by a separate map.
	  This reduces memory consumption by half.

Signed-off-by: Joe Tsai <joetsai@digital-static.net>